# src/main/java/com/api/transaction_aggregation/auth/dto/LoginRequest.java
package com.api.transaction_aggregation.auth.dto;

import jakarta.validation.constraints.NotBlank;

public record LoginRequest(
        @NotBlank(message = "Username is required")
        String username,

        @NotBlank(message = "Password is required")
        String password
) {}